requirements.txt
absl-py==1.3.0
anndata==0.8.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
autograd==1.5
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chex==0.1.5
click==8.1.3
cloudpickle==2.2.0
comm==0.1.2
commonmark==0.9.1
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
diffrax==0.2.2
dill==0.3.6
distrax==0.1.3
dm-tree==0.1.7
docker-pycreds==0.4.0
entrypoints==0.4
equinox==0.9.2
etils==0.9.0
executing==1.2.0
fastjsonschema==2.16.2
flatbuffers==22.11.23
flax==0.6.7
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
gast==0.4.0
-e git+https://github.com/oxcsml/geomstats.git@657a81674d993d77aac3958d6ebc69950211306f#egg=geomstats
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
h5py==3.7.0
idna==3.4
immutabledict==2.2.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
jax==0.4.8
jaxlib==0.4.7+cuda11.cudnn86
jaxtyping==0.2.9
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
keras==2.11.0
kfac-jax @ git+https://github.com/deepmind/kfac-jax@2a361b24ac28d39f7eaac72b5444565b44fc4e82
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
ml-collections==0.1.1
ml-dtypes==0.0.4
msgpack==1.0.4
munch==2.5.0
natsort==8.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvcc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==8.9.0.131
nvidia-cufft-cu11==10.9.0.58
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.1.5
orbax==0.1.6
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.13.1
POT==0.8.2
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
pyscf==2.1.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
Send2Trash==1.8.0
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tenacity==8.2.2
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.11.0
tensorflow-gan==2.1.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.28.0
tensorflow-metadata==1.12.0
tensorflow-probability==0.19.0
tensorstore==0.1.28
termcolor==2.1.1
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typeguard==2.13.3
typing_extensions==4.4.0
uri-template==1.2.0
urllib3==1.26.13
wandb==0.13.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.11.0
